Refactor(useWebSocket.tsx) : websocket 관련 작업 custom hook으로 분리
